fix(element): :bug: heading tracker no active heading
fix #26

bug: heading tracker should still indicate atleast one active heading.

fix: add state to element to indicate that it's the last visible element
to leave the viewport. the element should have been invisible, but we
mark it as still "visible", by not removing it from `activeHeadings` and
give it `data-last-visible` tag. then, on intersecting event, it will
been handled dynamically.